test thumbnail management of AlbumEditForm
